# Execute a single db transaction for all add operations