@notoco 2020-10-14
The big difference is also felt on the s905x2. Good job